Global SM with Player SM
Created player movement state machine that has idle, walk, and jump behaviors as the states. Minigame state machine uses events to send a broadcast to player state machine and player state machine uses that broadcast to decide whether or not to allow the player SM to be active. Also made the MinigameStateMachine working and switches between scenes. Still need to add more minigames/scenes for that SM.