ALMOST GAME FINISH
Lennon and I got the player switching working. Now when the spline is cleared, it picks a random minigame and sends the player to that. If the player's boardPosition is equal to whatever SplineGen winTile is, then it sets a bool in PlayerData to true which tracks if the player has completed the game or not. Then when the player is switched with GameManager, it checks if that bool for the currentPlayer is true or not. If it is true, then it just calls the NextPlayer method again.
Also got the destination fixed so now the destination is not constant over the various turns.
Most just need to add the cards and print the cards.